# api/core.openapi.yaml
basePath: /
definitions:
...
  TraceEvent:
    properties:
      created_at:
        type: string
      description:
        type: string
      entity_id:
        type: string
      entity_type:
        type: string
      event_type:
        type: string
      parent_span_id:
        type: string
      source:
        type: string
      span_id:
        type: string
      status:
        type: string
    type: object
  TraceNode:
    properties:
      children:
        items:
          $ref: '#/definitions/TraceNode'
        type: array
      created_at:
        type: string
      description:
        type: string
      entity_id:
        type: string
      entity_type:
        type: string
      event_type:
        type: string
      parent_span_id:
        type: string
      source:
        type: string
      span_id:
        type: string
      status:
        type: string
    type: object
  TraceSummary:
    properties:
      ended_at:
        type: string
      event_count:
        type: integer
      root_operation:
        type: string
      source:
        type: string
      started_at:
        type: string
      status:
        type: string
      trace_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'REST API for Apache Custos: organizations, users, projects, compute
    clusters, allocations, audit traces, and the privilege/role authorization layer.
    OIDC bearer auth required. Connector endpoints live under /connectors/<name>/.'
  title: Apache Custos Core API
  version: 0.2.0
paths:
  /audit/events:
    get:
      parameters:
      - description: 32-char lowercase hex
        in: query
        name: trace_id
        required: true
        type: string
      - description: 16-char lowercase hex
        in: query
        name: span_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              events:
                items:
                  $ref: '#/definitions/TraceEvent'
                type: array
            type: object
        "400":
          description: Malformed trace_id or span_id
          schema:
            properties:
              error:
                type: string
            type: object
        "503":
          description: Audit trace store not configured
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List audit events for a trace (optionally one span)
      tags:
      - Audit
  /audit/sources:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              sources:
                items:
                  type: string
                type: array
            type: object
        "503":
          description: Audit trace store not configured
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List distinct audit-event sources
      tags:
      - Audit
  /audit/traces:
    get:
      description: One row per distinct trace_id. `q` matches trace_id hex prefix
        or root-operation substring.
      parameters:
      - collectionFormat: multi
        description: Restrict to listed sources (repeatable)
        in: query
        items:
          type: string
        name: source
        type: array
      - collectionFormat: multi
        description: ok | error | in_progress (repeatable)
        in: query
        items:
          type: string
        name: status
        type: array
      - description: RFC3339 lower bound on started_at
        in: query
        name: from
        type: string
      - description: RFC3339 upper bound on started_at
        in: query
        name: to
        type: string
      - description: Substring match on trace_id hex or root operation
        in: query
        name: q
        type: string
      - description: Page size (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              limit:
                type: integer
              offset:
                type: integer
              total:
                type: integer
              traces:
                items:
                  $ref: '#/definitions/TraceSummary'
                type: array
            type: object
        "400":
          description: Invalid filter value
          schema:
            properties:
              error:
                type: string
            type: object
        "503":
          description: Audit trace store not configured
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List audit traces with filters
      tags:
      - Audit
  /audit/traces/{trace_id}:
    get:
      parameters:
      - description: 32-char lowercase hex
        in: path
        name: trace_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              trace_id:
                type: string
              tree:
                items:
                  $ref: '#/definitions/TraceNode'
                type: array
              truncated:
                type: boolean
            type: object
        "400":
          description: Malformed trace_id
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Trace not found
          schema:
            properties:
              error:
                type: string
            type: object
        "503":
          description: Audit trace store not configured
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a trace's nested span tree
      tags:
      - Audit
  /compute-allocation-change-request-events:
    post:
      consumes:
      - application/json
      parameters:
      - description: Event payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ComputeAllocationChangeRequestEvent'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ComputeAllocationChangeRequestEvent'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a change request event
      tags:
      - Change Request Events
  /compute-allocation-change-request-events/{id}:
    delete:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a change request event
      tags:
      - Change Request Events
    get:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ComputeAllocationChangeRequestEvent'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a change request event
      tags:
      - Change Request Events
  /compute-allocation-change-requests:
    post:
      consumes:
      - application/json
      parameters:
      - description: Change request payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ComputeAllocationChangeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ComputeAllocationChangeRequest'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a compute allocation change request
      tags:
      - Compute Allocation Change Requests
  /compute-allocation-change-requests/{id}:
    delete:
      parameters:
      - description: Change request ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a compute allocation change request
      tags:
      - Compute Allocation Change Requests
    get:
      parameters:
      - description: Change request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ComputeAllocationChangeRequest'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a compute allocation change request
      tags:
      - Compute Allocation Change Requests
    put:
      consumes:
      - application/json
      parameters:
      - description: Change request ID
        in: path
        name: id
        required: true
        type: string
      - description: Change request payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ComputeAllocationChangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ComputeAllocationChangeRequest'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a compute allocation change request
      tags:
      - Compute Allocation Change Requests
  /compute-allocation-change-requests/{id}/events:
    get:
      parameters:
      - description: Change request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ComputeAllocationChangeRequestEvent'
            type: array
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List events for a change request
      tags:
      - Change Request Events
  /compute-allocation-change-requests/{id}/events/latest:
    get:
      parameters:
      - description: Change request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ComputeAllocationChangeRequestEvent'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get the most recent event for a change request
      tags:
      - Change Request Events
  /compute-allocation-diffs:
    post:
      consumes:
      - application/json
      description: Records a discrete change against a compute allocation (e.g. USAGE_UPDATE
        or ALLOCATION_STATUS_CHANGE).
      parameters:
      - description: Diff payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ComputeAllocationDiff'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ComputeAllocationDiff'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a compute allocation diff
      tags:
      - Compute Allocation Diffs
  /compute-allocation-diffs/{id}:
    delete:
      parameters:
      - description: Diff ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a compute allocation diff
      tags:
      - Compute Allocation Diffs
    get:
      parameters:
      - description: Diff ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ComputeAllocationDiff'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a compute allocation diff
      tags:
      - Compute Allocation Diffs
  /compute-allocation-membership-resource-overrides:
    post:
      consumes:
      - application/json
      parameters:
      - description: Override payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ComputeAllocationMembershipResourceOverride'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ComputeAllocationMembershipResourceOverride'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a membership resource override
      tags:
      - Membership Resource Overrides
  /compute-allocation-membership-resource-overrides/{id}:
    delete:
      parameters:
      - description: Override ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a membership resource override
      tags:
      - Membership Resource Overrides
    get:
      parameters:
      - description: Override ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ComputeAllocationMembershipResourceOverride'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a membership resource override
      tags:
      - Membership Resource Overrides
    put:
      consumes:
      - application/json
      parameters:
      - description: Override ID
        in: path
        name: id
        required: true
        type: string
      - description: Override payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ComputeAllocationMembershipResourceOverride'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ComputeAllocationMembershipResourceOverride'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a membership resource override
      tags:
      - Membership Resource Overrides
  /compute-allocation-memberships:
    post:
      consumes:
      - application/json
      parameters:
      - description: Membership payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ComputeAllocationMembership'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ComputeAllocationMembership'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a compute allocation membership
      tags:
      - Compute Allocation Memberships
...
  /compute-allocation-memberships/{id}/resource-overrides:
    get:
      parameters:
      - description: Membership ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ComputeAllocationMembershipResourceOverride'
            type: array
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List resource overrides for a membership
      tags:
      - Membership Resource Overrides
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"